:truck: rename latam files schema
