+++ scrapers/latam_scraper/utils_latam.py
